#122: support for 2Gb+ files

+++ library/src/main/java/com/danikula/videocache/ByteArrayCache.java
@@ -33,7 +33,7 @@ public class ByteArrayCache implements Cache {
     }
 
     @Override
-    public int available() throws ProxyCacheException {
+    public long available() throws ProxyCacheException {
         return data.length;
     }
