+++ library/src/main/java/com/danikula/videocache/Source.java
@@ -13,7 +13,7 @@ public interface Source {
      * @param offset offset in bytes for source.
      * @throws ProxyCacheException if error occur while opening source.
      */
-    void open(int offset) throws ProxyCacheException;
+    void open(long offset) throws ProxyCacheException;
 
     /**
      * Returns length bytes or <b>negative value</b> if length is unknown.
@@ -21,7 +21,7 @@ public interface Source {
      * @return bytes length
      * @throws ProxyCacheException if error occur while fetching source data.
      */
-    int length() throws ProxyCacheException;
+    long length() throws ProxyCacheException;
 
     /**
      * Read data to byte buffer from source with current offset.
