+++ library/src/main/java/com/danikula/videocache/SourceInfo.java
@@ -8,10 +8,10 @@ package com.danikula.videocache;
 public class SourceInfo {
 
     public final String url;
-    public final int length;
+    public final long length;
     public final String mime;
 
-    public SourceInfo(String url, int length, String mime) {
+    public SourceInfo(String url, long length, String mime) {
         this.url = url;
         this.length = length;
         this.mime = mime;
