+++ library/src/main/java/com/danikula/videocache/file/FileCache.java
@@ -41,7 +41,7 @@ public class FileCache implements Cache {
     }
 
     @Override
-    public synchronized int available() throws ProxyCacheException {
+    public synchronized long available() throws ProxyCacheException {
         try {
             return (int) dataFile.length();
         } catch (IOException e) {
